pyproject.toml
[project]
name = "test-mlflow-remote"
version = "0.1.0"
description = "Testing Databricks-managed MLflow from a local IDE"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core MLflow and Databricks dependencies
    "mlflow[databricks]>=3.0.0,<4.0.0",
    "databricks-sdk>=0.18.0",
    "databricks-cli>=0.18.0",
    # For Databricks Connect (optional, for running Spark jobs)
    "databricks-connect>=15.0.0",
    # For GenAI agent testing (optional)
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    # Utilities
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    # For testing
    "pytest>=7.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.uv.sources]

